refactor(records): reading a relationship edge moves to its own file
relationship.go had grown back past the 500-LOC cap the gate enforces.
The reads and the row-scope probes that guard them are one concept and
they leave together: visibleRelationship, GetRelationship, the two
EnsureVisible probes and the column aliaser. What stays is the write
path and the write shape it lands.

Verbatim move — no behaviour changes.
